src/requirements/requirements.txt
smplx[all]
numpy>=1.16.2
torch>=1.0.1.post2
dataclasses>=0.6
ruamel.yaml
chumpy